Extract common StytchClient code to internal protocol
