+++ Sources/StytchCore/StytchClient/StytchClient.swift
@@ -8,25 +8,13 @@ import Foundation
  **Async Options**: Async function calls for Stytch products are available via various
  mechanisms (Async/Await, Combine, callbacks) so you can use whatever best suits your needs.
  */
-public struct StytchClient {
+public struct StytchClient: StytchClientType {
     static var instance: StytchClient = .init()
 
-    static var router: NetworkingRouter<BaseRoute> = .init()
-
-    private(set) var configuration: Configuration? = {
-        guard
-            let url = Bundle.main.url(forResource: "StytchConfiguration", withExtension: "plist"),
-            let data = try? Data(contentsOf: url)
-        else { return nil }
-
-        return try? PropertyListDecoder().decode(Configuration.self, from: data)
-    }() {
-        didSet { updateHeaderProvider() }
-    }
+    static var router: NetworkingRouter<BaseRoute> = .init { instance.configuration }
 
     private init() {
-        updateHeaderProvider()
-        runKeychainMigrations()
+        postInit()
     }
 
     /**
@@ -36,7 +24,12 @@ public struct StytchClient {
        - hostUrl: Generally this is your backend's base url, where your apple-app-site-association file is hosted. This is an https url which will be used as the domain for setting session-token cookies to be sent to your servers on subsequent requests. If not passed here, no cookies will be set on your behalf.
      */
     public static func configure(publicToken: String, hostUrl: URL? = nil) {
-        instance.configuration = .init(hostUrl: hostUrl, publicToken: publicToken)
+        _configure(publicToken: publicToken, hostUrl: hostUrl)
+    }
+
+    ///  A helper function for parsing out the Stytch token types and values from a given deeplink
+    public static func tokenValues(for url: URL) throws -> (DeeplinkTokenType, String)? {
+        try _tokenValues(for: url)
     }
 
     // sourcery: AsyncVariants, (NOTE: - must use /// doc comment styling)
@@ -50,7 +43,7 @@ public struct StytchClient {
     public static func handle(
         url: URL,
         sessionDuration: Minutes = .defaultSessionDuration
-    ) async throws -> DeeplinkHandledStatus<AuthenticateResponseType> {
+    ) async throws -> DeeplinkHandledStatus<AuthenticateResponse> {
         guard let (tokenType, token) = try tokenValues(for: url) else {
             return .notHandled
         }
@@ -62,55 +55,6 @@ public struct StytchClient {
             return try await .handled(oauth.authenticate(parameters: .init(token: token, sessionDuration: sessionDuration)))
         case .passwordReset:
             return .manualHandlingRequired(.passwordReset, token: token)
-        }
-    }
-
-    ///  A helper function for parsing out the Stytch token types and values from a given deeplink
-    public static func tokenValues(for url: URL) throws -> (DeeplinkTokenType, String)? {
-        guard
-            let components = URLComponents(url: url, resolvingAgainstBaseURL: true),
-            let queryItems = components.queryItems,
-            let typeQuery = queryItems.first(where: { $0.name == "stytch_token_type" }), let type = typeQuery.value,
-            let tokenQuery = queryItems.first(where: { $0.name == "token" }), let token = tokenQuery.value
-        else {
-            return nil
-        }
-        guard let tokenType = DeeplinkTokenType(rawValue: type) else {
-            throw StytchError.unrecognizedDeeplinkTokenType
-        }
-        return (tokenType, token)
-    }
-
-    // To be called after configuration
-    private func updateHeaderProvider() {
-        let clientInfoString = try? Current.clientInfo.base64EncodedString()
-
-        Current.networkingClient.headerProvider = {
-            guard let configuration = Self.instance.configuration else { return [:] }
-
-            let sessionToken = Current.sessionStorage.sessionToken?.value ?? configuration.publicToken
-            let authToken = "\(configuration.publicToken):\(sessionToken)".base64Encoded()
-
-            return [
-                "Content-Type": "application/json",
-                "Authorization": "Basic \(authToken)",
-                "X-SDK-Client": clientInfoString ?? "",
-            ]
-        }
-    }
-
-    private func runKeychainMigrations() {
-        KeychainClient.migrations.forEach { migration in
-            let migrationName = "stytch_keychain_migration_" + String(describing: migration.self)
-            guard !Current.defaults.bool(forKey: migrationName) else {
-                return
-            }
-            do {
-                try migration.run()
-                Current.defaults.set(true, forKey: migrationName)
-            } catch {
-                print(error)
-            }
         }
     }
 }
