--- Sources/StytchCore/StytchClient/User/LocalStorage+User.swift
@@ -1,10 +1,0 @@
-struct UserStorageKey: LocalStorageKey {
-    typealias Value = User
-}
-
-extension LocalStorage {
-    var user: User? {
-        get { Current.localStorage[UserStorageKey.self] }
-        set { Current.localStorage[UserStorageKey.self] = newValue }
-    }
-}
